SpellbookFragment update

+++ android/app/src/main/java/com/example/yourrpg/model/Character.java
@@ -96,7 +96,7 @@ public class Character implements Serializable {
         this.level = level;
     }
 
-    public double getExp() {
+    public int getExp() {
         return exp;
     }
